added layer abstraction, changed output type of BMat matmul to float

+++ include/bdlearn/BMat.hpp
@@ -7,7 +7,6 @@
 #include "Halide.h"
 
 namespace bdlearn {
-    // All rows of BMats are byte-aligned, so each row is zero-padded out on the right
     class BMat {
         public:
         // Constructors
@@ -30,7 +29,7 @@ namespace bdlearn {
         friend bool operator==(const BMat& a, const BMat& b);
         friend std::ostream& operator<<(std::ostream& os, const BMat& bmat);
         // C += A @ B
-        friend void matmul(Halide::Buffer<int>* dest, const BMat& A, const BMat& B);
+        friend void matmul(Halide::Buffer<float>* dest, const BMat& A, const BMat& B);
 
 
         private:
